fix: minor refactor
